Cast token count from size_t to int in C#

+++ test/csharp/TestOnnxRuntimeGenAIAPI.cs
@@ -163,7 +163,7 @@ namespace Microsoft.ML.OnnxRuntimeGenAI.Tests
                             Assert.False(generator.IsDone());
                             Assert.Equal(generatorParams.GetSearchNumber("max_length"), maxLength);
                             Assert.Equal(generatorParams.GetSearchBool("early_stopping"), true);
-                            Assert.Equal(generator.TokenCount(), generator.GetSequence(0).Length);
+                            Assert.Equal((int)generator.TokenCount(), generator.GetSequence(0).Length);
 
                             while (!generator.IsDone())
                             {
@@ -175,7 +175,7 @@ namespace Microsoft.ML.OnnxRuntimeGenAI.Tests
                                 var sequence = generator.GetSequence(i).ToArray();
                                 var expectedSequence = expectedOutput.Skip((int)i * (int)maxLength).Take((int)maxLength);
                                 Assert.Equal(expectedSequence, sequence);
-                                Assert.Equal(generator.TokenCount(), generator.GetSequence(i).Length);
+                                Assert.Equal((int)generator.TokenCount(), generator.GetSequence(i).Length);
                             }
                         }
                     }
